requirements.txt
# Core ML and NLP libraries
torch>=1.9.0
transformers>=4.21.0
sentence-transformers>=2.2.0
datasets>=2.0.0
tokenizers>=0.13.0

# Data processing and analysis
pandas>=1.3.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0

# Vector search and similarity
faiss-cpu>=1.7.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Web app framework
streamlit>=1.28.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0

# For creating flowcharts (optional)
graphviz>=0.20.0